Removed non-CMD version (CMD is vastly superior)
