--- Assets/Materials/BlueGlowObj.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5c7655b2e2a47834986bd427d3d870ef
-timeCreated: 1484542383
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
