--- Assets/Materials/CmdBufferGlow.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f1c6a12aebf887a4d959a64ea35106d8
-timeCreated: 1486111862
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
