--- Assets/Materials/GreenGlowObj.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 66443556d664e014894c68cda9295d67
-timeCreated: 1484541624
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
