--- Assets/Materials/RedGlowObj.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7927fc1ca4e84e142a6f8a9bf2ed3e7f
-timeCreated: 1484541624
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
