+++ Assets/Materials/StrongBlue.mat
@@ -6,7 +6,7 @@ Material:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
-  m_Name: RedCmdGlowObj
+  m_Name: StrongBlue
   m_Shader: {fileID: 46, guid: 0000000000000000f000000000000000, type: 0}
   m_ShaderKeywords: 
   m_LightmapFlags: 4
@@ -72,5 +72,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.7372549, g: 0.043137256, b: 0.043137256, a: 1}
-    - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
+    - _Color: {r: 0.043137256, g: 0.043137256, b: 0.7372549, a: 1}
+    - _EmissionColor: {r: 0.022707596, g: 0.025131233, b: 0.03676468, a: 1}
