+++ Assets/Materials/StrongBlue.mat.meta
