+++ Assets/Materials/StrongGreen.mat
@@ -6,10 +6,10 @@ Material:
   m_ObjectHideFlags: 0
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 0}
-  m_Name: RedGlowObj
-  m_Shader: {fileID: 4800000, guid: 442f40f05878ff645995c0d278686da4, type: 3}
-  m_ShaderKeywords: _EMISSION
-  m_LightmapFlags: 1
+  m_Name: StrongGreen
+  m_Shader: {fileID: 46, guid: 0000000000000000f000000000000000, type: 0}
+  m_ShaderKeywords: 
+  m_LightmapFlags: 4
   m_EnableInstancingVariants: 0
   m_DoubleSidedGI: 0
   m_CustomRenderQueue: -1
@@ -72,5 +72,5 @@ Material:
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
-    - _Color: {r: 0.7372549, g: 0.043137256, b: 0.043137256, a: 1}
-    - _EmissionColor: {r: 0.04411763, g: 0.033412617, b: 0.033412617, a: 1}
+    - _Color: {r: 0.043137256, g: 0.7372549, b: 0.043137256, a: 1}
+    - _EmissionColor: {r: 0, g: 0, b: 0, a: 1}
