+++ Assets/Materials/StrongGreen.mat.meta
