+++ Assets/Materials/StrongRed.mat.meta
