--- Assets/Materials/WhiteGlowObj.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 59b10b2fe6aef3a47bc0bb5646406cfe
-timeCreated: 1484542383
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
