+++ Assets/Resources/GlowComposite.shader
@@ -7,7 +7,7 @@
 
 		Pass {
 			CGPROGRAM
-			#pragma vertex vert_img //Built-in declaration of appdata and v2f, sending just position and uv (texture coordinate)
+			#pragma vertex vert_img //Built-in declaration of appdata and v2f for image effects, sending just position and uv (texture coordinate)
 			#pragma fragment frag
 			#include "UnityCG.cginc"
 			
@@ -16,14 +16,14 @@
 			sampler2D _GlowBlurredTex;
 			float _Intensity;
 
-			fixed4 frag (v2f_img IN) : SV_Target {
-				fixed4 col = tex2D(_MainTex, IN.uv);
+			half4 frag (v2f_img IN) : SV_Target {
+				half4 col = tex2D(_MainTex, IN.uv);
 
 				#if !UNITY_UV_STARTS_AT_TOP //Invert Y in case of D3D?
 	      			IN.uv.y = 1 - IN.uv.y; //Unnecessary? if (_MainTex_TexelSize.y < 0)
 				#endif
 
-				fixed4 glow = max(0, tex2D(_GlowBlurredTex, IN.uv) - tex2D(_GlowPrePassTex, IN.uv));
+				half4 glow = max(0, tex2D(_GlowBlurredTex, IN.uv) - tex2D(_GlowPrePassTex, IN.uv));
 				return col + glow * _Intensity;
 			}
 			ENDCG
