--- Assets/Resources/GlowReplace.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 45f4b5fa143a31c4a96299642e84e286
-timeCreated: 1484547336
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
