--- Assets/Resources/Standard-Glow.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 442f40f05878ff645995c0d278686da4
-timeCreated: 1484547038
-licenseType: Pro
-ShaderImporter:
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
