--- Assets/Scenes/cmd_buffer_variant.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 987d53ff511d51f479bfaf1443522ceb
-timeCreated: 1484541434
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
