--- Assets/Scripts/CMD_Buffer_Scripts.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: bb85618ee84fbff46b61b8f592cf9437
-folderAsset: yes
-timeCreated: 1530992250
-licenseType: Free
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
