--- Assets/Scripts/CMD_Buffer_Scripts/GlowObjectCmd.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8f97577f7490be846abb026b83b52db0
-timeCreated: 1484547418
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
