--- Assets/Scripts/GlowComposite.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d4c072b0c78ba8a4ebec1edd40d3eff2
-timeCreated: 1484548498
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
