--- Assets/Scripts/GlowPrePass.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: a8d6d491c430e2d4ea283009ec2f2794
-timeCreated: 1484547290
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
